최신 버전 반영

+++ Pinokio.Simulator/ModelDesigner/ModelDesigner_NodeModify.cs
@@ -92,11 +92,12 @@ namespace Pinokio.Designer
 
         private void DeletePartReference(PartReference EntityReference)
         {
-            partTreeList.BeginInvoke(new Action(() =>
-            {
-                PartTreeListNode treeNode = (PartTreeListNode)partTreeList.FindNodeByKeyID(EntityReference.ID);
-                partTreeList.DeleteNode(treeNode);
-            }));
+            if (ModelManager.Instance.AnimationNode != null && ModelManager.Instance.AnimationNode.IsUse)
+                partTreeList.BeginInvoke(new Action(() =>
+                {
+                    PartTreeListNode treeNode = (PartTreeListNode)partTreeList.FindNodeByKeyID(EntityReference.ID);
+                    partTreeList.DeleteNode(treeNode);
+                }));
         }
 
         private void CutNode()
